+++ samples/tab-corporate-helper/Controllers/ExpenseController.cs
@@ -1,25 +1,30 @@
-﻿using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Configuration;
-using System.Diagnostics;
-using System.Net.Http;
-using TeamsAuthSSO.Models;
+﻿using Microsoft.AspNetCore.Mvc;
+using System.Reflection;
 
-namespace CoorporateHelper.Controllers
+namespace CorporateHelper.Controllers
 {
-	public class VacationController : Controller
+
+	using System.Diagnostics;
+	using System.Net.Http;
+	using Microsoft.AspNetCore.Http;
+	using Microsoft.AspNetCore.Mvc;
+	using Microsoft.Extensions.Configuration;
+	using TeamsAuthSSO.Models;
+
+	public class ExpenseController : Controller
 	{
+
 		private readonly IConfiguration _configuration;
 		private readonly IHttpClientFactory _httpClientFactory;
 		private readonly IHttpContextAccessor _httpContextAccessor;
 
 		/// <summary>
-		/// Initializes a new instance of the <see cref="VacationController"/> class.
+		/// Initializes a new instance of the <see cref="ExpenseController"/> class.
 		/// </summary>
 		/// <param name="configuration">IConfiguration instance.</param>
 		/// <param name="httpClientFactory">IHttpClientFactory instance.</param>
 		/// <param name="httpContextAccessor">IHttpContextAccessor instance.</param>
-		public VacationController(
+		public ExpenseController(
 			IConfiguration configuration,
 			IHttpClientFactory httpClientFactory,
 			IHttpContextAccessor httpContextAccessor)
@@ -29,7 +34,7 @@ namespace CoorporateHelper.Controllers
 			_httpContextAccessor = httpContextAccessor;
 		}
 
-		public IActionResult Vacations()
+		public IActionResult Expenses()
 		{
 			return View();
 		}
