+++ samples/tab-corporate-helper/Helper/SSOAuthHelper.cs
