+++ samples/tab-corporate-helper/Models/ErrorViewModel.cs
