+++ samples/tab-corporate-helper/Properties/launchSettings.json
