+++ samples/tab-corporate-helper/Views/Home/Index.cshtml
@@ -1,8 +1,11 @@
 ﻿@{
-    ViewData["Title"] = "Coorporate Helper";
+    ViewData["Title"] = "Corporate Helper";
 }
-<h1>Coorporate Helper</h1>
+<h1>Corporate Helper</h1>
 <p>Welcome to your one-place for submitting your cooporate needs! Please check the tabs above for the tools that you need.</p>
 <div id="divError" style="display: none"></div>
 <button onclick="requestConsent()" id="consent" style="display:none;">Authenticate</button>
 <div id="divGraphProfile" style="display: none"></div>
+<div id="divOutsideTeams" style="display: none">
+    <p>You have opened this page outside of Microsoft Teams! Please install this app as a Teams App.</p>
+</div>
