+++ samples/tab-corporate-helper/appPackage/color.png
