+++ samples/tab-corporate-helper/appPackage/manifest.json
@@ -3,20 +3,20 @@
   "manifestVersion": "1.12",
   "version": "1.0.5",
   "id": "<<YOUR-MICROSOFT-APP-ID>>",
-  "packageName": "com.microsoft.coorporatehelper",
+  "packageName": "com.microsoft.Corporatehelper",
   "developer": {
-    "name": "Felipe Dotti",
+    "name": "Microsoft",
     "websiteUrl": "https://www.microsoft.com",
     "privacyUrl": "https://www.microsoft.com/privacy",
     "termsOfUseUrl": "https://www.microsoft.com/termsofuse"
   },
   "name": {
-    "short": "Coorporate Helper",
-    "full": "Coorporate Helper"
+    "short": "Corporate Helper",
+    "full": "Corporate Helper"
   },
   "description": {
-    "short": "Coorporate Helper",
-    "full": "Coorporate Helper"
+    "short": "Corporate Helper",
+    "full": "Corporate Helper"
   },
   "icons": {
     "outline": "outline.png",
