+++ samples/tab-corporate-helper/appPackage/outline.png
