+++ samples/tab-corporate-helper/wwwroot/favicon.ico
